BaseAudioSharedUnit::devicesChanged is failing capture too much in GPUProcess https://bugs.webkit.org/show_bug.cgi?id=235288
Reviewed by Eric Carlson.

Manually tested.

* platform/mediastream/mac/CoreAudioCaptureDeviceManager.h:
In GPUProcess, we do not need to filter tap enabled devices, since this may filter out valid devices, which may break capture.
Given we only need to filter tap enabled devices in the process that enumerates devices for selecting the one to use,
We can reverse the default value to not filter, and enable the filtering in UIProcess as done currently based on whether capturing in GPUProcess or not.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288110 268f45cc-cd09-0410-ab3c-d52691b4dbfc