Improve computation of service worker FetchEvent.resultingClientId https://bugs.webkit.org/show_bug.cgi?id=235107
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

* web-platform-tests/service-workers/service-worker/clients-get-resultingClientId.https-expected.txt:
* web-platform-tests/service-workers/service-worker/clients-get.https-expected.txt:
* web-platform-tests/service-workers/service-worker/historical.https.any.serviceworker-expected.txt:

Source/WebCore:

We are now passing the reserved client ID generated by DocumentLoader to the generated document.
Previously, we would register a temporary ID that we would drop later on when creating a new document.
This allows to be closer to spec, though we need to handle the case of a document whose origin gets unique
due to CSP or sandboxing making a document origin unique.
We are also removing targetClientId, which was renamed in the spec to replacesClientId but is also not implemented in any browser yet.

Covered by existing tests.

* dom/DOMImplementation.cpp:
* dom/DOMImplementation.h:
* dom/Document.cpp:
* dom/Document.h:
* dom/ScriptExecutionContext.cpp:
* dom/ScriptExecutionContext.h:
* html/FTPDirectoryDocument.cpp:
* html/HTMLDocument.cpp:
* html/HTMLDocument.h:
* html/ImageDocument.cpp:
* html/MediaDocument.cpp:
* html/ModelDocument.cpp:
* html/PluginDocument.cpp:
* html/TextDocument.cpp:
* html/TextDocument.h:
* inspector/DOMPatchSupport.cpp:
* loader/DocumentLoader.cpp:
* loader/DocumentLoader.h:
* loader/DocumentWriter.cpp:
* loader/DocumentWriter.h:
* loader/SinkDocument.cpp:
* workers/service/FetchEvent.cpp:
* workers/service/FetchEvent.h:
* workers/service/FetchEvent.idl:
* workers/service/server/SWServer.cpp:
* xml/XMLHttpRequest.cpp:

Source/WebKit:

* WebProcess/Network/WebLoaderStrategy.cpp:
* WebProcess/Storage/WebSWClientConnection.cpp:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288201 268f45cc-cd09-0410-ab3c-d52691b4dbfc