[JSC] Relax Date.parse requirement https://bugs.webkit.org/show_bug.cgi?id=235468
Reviewed by Darin Adler.

JSTests:

* stress/date-relaxed-separator.js: Added.
(shouldBe):

Source/WTF:

While the spec does not require accepting 't' / ' ' separator, ISO 8601 accepts it.
This is because ECMA262's Date format is *not* ISO 8601 (it is called simplification
of ISO 8601[1]).
This patch relaxes this strictness to accept more formats, which can be accepted in
the other engines too.

[1]: https://tc39.es/ecma262/#sec-date-time-string-format

* wtf/DateMath.cpp:
(WTF::parseES5DateFromNullTerminatedCharacters):

LayoutTests:

* js/date-parse-test-expected.txt:
* js/script-tests/date-parse-test.js:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288411 268f45cc-cd09-0410-ab3c-d52691b4dbfc