[LFC][IFC] InlineFormattingContext::computeGeometryForLineContent should always return the line logical rect https://bugs.webkit.org/show_bug.cgi?id=235494
Reviewed by Antti Koivisto.

The return rect of computeGeometryForLineContent (based on the display line) is input to the next line's
initial rect and therefore it is supposed to be logical (in general, layout should strictly use logical coords).

This is in preparation for supporting vertical writing mode.

* layout/formattingContexts/inline/InlineFormattingContext.cpp:
(WebCore::Layout::InlineFormattingContext::computeGeometryForLineContent):
* layout/formattingContexts/inline/InlineFormattingGeometry.cpp:
(WebCore::Layout::InlineFormattingGeometry::flipVisualRectToLogicalForWritingMode):
* layout/formattingContexts/inline/InlineFormattingGeometry.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288421 268f45cc-cd09-0410-ab3c-d52691b4dbfc