Position:fixed layers shouldn't allocate a backing buffer if all children are offscreen. https://bugs.webkit.org/show_bug.cgi?id=235420 <rdar://86612099>
Patch by Matt Woodrow <mattwoodrow@apple.com> on 2022-01-23
Reviewed by Simon Fraser and Darin Adler.

Source/WebCore:

Adds a bounds intersection check to isPaintDestinationForDescendantLayers,
so that we can exclude descendants that definitely won't draw anything into
the compositing layer. Uses a conservative check, which gives up if there are
any transforms in the ancestor chain.

Test: compositing/backing/no-backing-for-offscreen-children-of-position-fixed.html

* rendering/RenderLayerBacking.cpp:
(WebCore::intersectsWithAncestor):
(WebCore::RenderLayerBacking::isPaintDestinationForDescendantLayers const):

LayoutTests:

Adds a test that has a viewport sized position:fixed element (with compositing
layer), and a single child which is entirely offscreen. Tests that we correctly
determine that we don't need a backing store for the layer.

* compositing/backing/no-backing-for-offscreen-children-of-position-fixed-expected.txt: Added.
* compositing/backing/no-backing-for-offscreen-children-of-position-fixed.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288429 268f45cc-cd09-0410-ab3c-d52691b4dbfc