[macOS] Update Pasteboard::read to prioritize native representations over TIFF https://bugs.webkit.org/show_bug.cgi?id=190101 rdar://44879244
Reviewed by Ryosuke Niwa.

Source/WebCore:

Currently on macOS, when pasting image data from the pasteboard into richly editable elements, we attempt to
read the image data by checking a hard-coded list of types; in order, these are: TIFF, PDF, PNG, and finally
JPEG. Since AppKit automatically provides TIFF as an available type for compatibility when writing non-TIFF
image data to the pasteboard, this means that WebKit automatically transcodes all images to TIFF when pasting
images in editable areas.

This leads to unexpected behavior in Mail compose when using the continuity camera to insert images, since all
drawings and photos end up being inserted as (often much larger) TIFF files instead of JPEG and PNG,
respectively. To address this, we adjust `Pasteboard::read` to prioritize image formats in the hard-coded list
above that were written directly to the pasteboard by the copier (i.e. without transcoding) before reading from
the remaining types that were added to the pasteboard by the system (and thus require extra transcoding).

In the case where only PNG and JPEG data was written to the pasteboard by the copier, this means we'll read
images as PNG and JPEG data directly. If TIFF was explicitly written as an image format to the pasteboard, this
patch will preserve existing behavior by reading the image data as TIFF.

Note that this existing logic for reading images directly into web content is an existing privacy risk, since
AppKit's pasteboard code preserves all metadata properties when transcoding between image formats. In the
future, we should add a step here to strip out as much metadata as possible here, with an allow-list of metadata
properties that are essential to preserving image fidelity (e.g. orientation).

See: <https://webkit.org/b/235534> for more details.

Test: WKAttachmentTests.PastingPreservesImageFormat

* platform/mac/PasteboardMac.mm:
(WebCore::Pasteboard::read):

Refactor this logic so that we first try to read any natively written image formats that happen to be one of
("TIFF", "PDF", "PNG", "JPEG"), before trying to read any other types that would otherwise require transcoding.

Tools:

Add a cross-platform API test to verify that when using the client-side attachments, pasted image data of
varying formats (PDF, JPEG, PNG) are read in their original form. See WebCore/ChangeLog for more details. This
test currently passes on iOS but fails on macOS, where both the PNG and JPEG images are transcoded and read as
TIFF.

* TestWebKitAPI/Tests/WebKitCocoa/WKAttachmentTests.mm:
(testJPEGFileURL):
(testJPEGData):
(platformCopyPDF):
(platformCopyJPEG):
(TestWebKitAPI::TEST):

Drive-by fix: also remove a JavaScript evaluation in a nearby test that could only result in an exception (and
was probably accidentally left there while writing the test).


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288477 268f45cc-cd09-0410-ab3c-d52691b4dbfc