Symbols not always properly hidden when using WebKitAdditions to introduce new API https://bugs.webkit.org/show_bug.cgi?id=235674 <rdar://87999257>
Reviewed by Wenson Hsieh.

Surround APIs from WebKitAdditions with `#if ENABLE(API_WEBKIT_ADDITIONS)` instead of
`#if USE(APPLE_INTERNAL_SDK)` and leaving API_WEBKIT_ADDITIONS to be intentionally
undefined. This is because we never want to build the code inside this #if block.
Instead, those blocks are only meant to be replaced by the
replace-webkit-additions-includes.py with the code from WebKitAdditions.

* UIProcess/API/Cocoa/WKWebpagePreferences.h:
* mac/replace-webkit-additions-includes.py:
(main):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288658 268f45cc-cd09-0410-ab3c-d52691b4dbfc