Avoid setting and clearing :active state when dispatching synthetic click events when possible https://bugs.webkit.org/show_bug.cgi?id=235672 <rdar://problem/88095418>
Reviewed by Simon Fraser.

Source/WebCore:

Simulated click events are dispatched with two options:

- whether to send associated events mouseover, mouseup, mousedown
- whether to repaint the target element with its pressed look

We currently always set the element's :active state just after when we'd
send the mousedown event, and clear it just after that.

When we dispatch a simulated click event with neither of the above
options set, there's no way to observe the temporary :active state on
the element. We can skip it in that case.

We need to continue clearing clearing the :active state regardless,
because some callers have already set :active and are relying on
simulateClick to clear it.

This patch is a 0.3-0.4% improvement on Speedometer 2.

* dom/SimulatedClick.cpp:
(WebCore::simulateClick):

LayoutTests:

* platform/gtk/inspector/timeline/line-column-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288669 268f45cc-cd09-0410-ab3c-d52691b4dbfc