DocumentContext gives empty rects for blank lines. https://bugs.webkit.org/show_bug.cgi?id=235695
Reviewed by Wenson Hsieh.

Source/WebCore:

Test: DocumentEditingContext::RectsRequestInContentEditable

When requesting the document context for the end of a document, we would give empty rects for
blank lines. This was because when calculating the absoluteBoundingRect for the range, we
would use a normal unionRect call, which would use the basic unite call, which does not combine
empty rects. And empty rect is defined as a rect that does not have a non-zero hight and width.
Since we were calculating information about <br> lines, this was causing us to union a list of
a single rect, which was an empty rect. This left us with a zero rect after unioning. The solution
is to unionIfNonZero when calculating these rects for DocumentContext. I also took the opportunity
to make some of this code more parallel to other implementations.

* platform/graphics/GeometryUtilities.cpp:
(WebCore::unionRectIfNonZero):
* platform/graphics/GeometryUtilities.h:
* platform/graphics/IntRect.cpp:
(WebCore::IntRect::uniteIfNonZero):
* platform/graphics/IntRect.h:
(WebCore::IntRect::isZero const):

Source/WebKit:

When requesting the document context for the end of a document, we would give empty rects for
blank lines. This was because when calculating the absoluteBoundingRect for the range, we
would use a normal unionRect call, which would use the basic unite call, which does not combine
empty rects. And empty rect is defined as a rect that does not have a non-zero hight and width.
Since we were calculating information about <br> lines, this was causing us to union a list of
a single rect, which was an empty rect. This left us with a zero rect after unioning. The solution
is to unionIfNonZero when calculating these rects for DocumentContext. I also took the opportunity
to make some of this code more parallel to other implementations.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::requestDocumentEditingContext):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/DocumentEditingContext.mm:
(TEST): DocumentEditingContext::RectsRequestInContentEditable


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288721 268f45cc-cd09-0410-ab3c-d52691b4dbfc