Collapsed ranges return zero rects in DocumentContext https://bugs.webkit.org/show_bug.cgi?id=235844
Reviewed by Tim Horton.

Source/WebKit:

Test: DocumentEditingContext.RectsRequestInContentEditableWithDivBreaks

When getting the DocumentContext for a line, followed by a blank line, the blank link will
be represented by a collapsed range. This is a result of the emitCharacter call from representNodeOffsetZero.
This possibly a bug, but rather than try and fix this issue within TextIterator itself, we are
mitigating this issue by instead of using the collapsed range and the resulting non-existent character rect,
using the caret bounding rect. This gives us a more accurate bounding rect for situations like
this that result in collapsed ranges. And this would cover all other situations that might result in
a collapsed range from TextIterator.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::requestDocumentEditingContext):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/DocumentEditingContext.mm:
(TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@288849 268f45cc-cd09-0410-ab3c-d52691b4dbfc